comment and code style
